fehler gefunden, a getext und make optimized

+++ Blatt3/aufgabe1/aufgabe1.cpp
@@ -32,7 +32,7 @@ public:
   int getZustand(){
     return spin;
   }
-  int getEnergie(double H){
+  double getEnergie(double H){
     return (-spin*H);
   }
   void changeZustand(){
